Add saving of drafts to HTML5 localStorage
Add an abstract ReviewStorage "interface" for saving information about
in-progress reviews and an implementation of in terms of the
"Web Storage" spec as implemented in very recent browsers.

Hook that up in the web interface to auto-save on changes (with a
"Saving Draft..." indicator that is at least useful while debugging
the facility.)

The "Save" button is retitled to "Publish" to reduce confusion between
saving drafts and publishing to the bug.