Add the "function line" to the top of each hunk
The "function line" is the text after the:

 @@ -I,J +K,L @@

in a unified diff - by default with 'git diff' or 'diff -P' it's the previous
line starting with '^[\w$]' which for C will likely include the name of
the enclosing function. (Configurable with 'diff -F' or via .gitattributes.)

Extract that when parsing hunks and show it in the hunk header.