Be robust against munged comments
Bugzilla 3.4 currently has bug where it line-wraps (and strips trailing
whitespace) from comments in the XML output.

Try to robust against this by ignoring lines we don't know about and
treating '' the same as ' '. If the wrong thing gets wrapped, we can
get confused and get a blank comment, but it's better than having
the page stop loading entirely.