Switch to explicit Save/Cancel buttons
Rather than making comments be saved on blur, add explicit buttons
to Save or Cancel. Add a Delete button as well for existing comments.

Code is restructured to keep the current comment in a global variable
and identify the comment editor div as #commentEditor for easy access.

Generically provide <div class='clear'/> to get clear: both.