Add a Bugzilla extension
Add a Bugzilla extension to integrate Splinter closely into Bugzilla.
It adds:

 - A template to display Splinter with Bugzilla headers and footers
 - Hooks to add review links to review comments and to the attachments
   section of edit_bug.cgi
 - A section in the Bugzilla parameters for splinter.

Most of the content of index.html is moved into index.html.body, which
is included (by Makefile rules) into index.html and into the
splinter.html template.

config.js no longer exists as a file; in the proxy case, config.js is
generated on-the-fly by splinter_proxy.py. In the Bugzilla extension
case, the config.js content is emitted by the splinter.html template.

The format of the standalone version of Splinter has been changed to be
compatible with the Bugzilla-template version: the header is a single
line with "Title   Subtitle    Information" and the margin is now added
in a <div/> that encloses all the non-header content.