Add Splinter.publish_review XML-RPC method
Add an XML-RPC method to "publish a review" - this is pretty much like
Bug.add_comment(), but allows us to update the attachment status
simultaneously and only send out one email.

Use that, when the extension is available, to publish reviews.

Add xmlrpc.cgi to the list of URL's whitelisted for forwarding
in splinter_proxy.py, to allow using the proxy for developing against
client code an extension-enabled server.